requirements.txt
pyrogram
pypragyanx
tgcrypto
python-dotenv[cli]
python-dotenv
gitpython
psutil
py-cpuinfo 
telegraph
